fixed error in rate-limit handling

+++ src/cache/mod.rs
@@ -8,9 +8,9 @@ use crate::prelude::Message;
 
 #[derive(Debug)]
 pub struct RateLimitInfo {
-    pub limit: u64,
-    pub remaining: u64,
-    pub reset: u64,
+    pub limit: u32,
+    pub remaining: u32,
+    pub reset: f64,
 }
 
 lazy_static::lazy_static! {
